It seems that we have finished with the weather and everything is fine according to the tests, we’ll see in the future

+++ stay-go/src/main/java/com/staygo/enity/weather/Hourly.java
@@ -12,5 +12,5 @@ import java.util.List;
 @NoArgsConstructor
 public class Hourly {
     private List<String> time;
-    private List<Double> temperature_2m;
+    private List<Integer> temperature_2m;
 }
